backend/requirements.txt
fastapi
fastapi-throttling
sqlmodel
uvicorn[standard]
psycopg2-binary
python-dotenv
httpx
pyjwt
python-multipart
cryptography
pwdlib[argon2]